requirements.txt
flask==2.0.1
openai==1.0.0
google-api-python-client==2.0.2
google-auth-oauthlib==0.4.6
nltk==3.6.3
streamlit==1.8.0
google-auth-httplib2
python-dotenv